backend/requirements.txt
fastapi==0.135.2
sentry-sdk[fastapi]==2.27.0
starlette==0.52.1
uvicorn[standard]==0.34.0
pydantic==2.11.3
pytest==9.0.3
httpx==0.28.1
slowapi==0.1.9
limits==3.14.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
aiosqlite==0.20.0
alembic==1.14.0
PyJWT==2.10.1
cryptography==44.0.2
